added validation middleware for POST and PATCH requests

+++ controllers/users.js
@@ -1,26 +1,25 @@
 const { StatusCodes } = require("http-status-codes");
-const { validateReqBodyFields } = require("../utils/helpers");
 const catchAsync = require("../utils/catchAsync");
-const { ALLOWED_USER_FIELDS } = require("../constants");
 const User = require("../models/User");
+const { ForbiddenError } = require("../utils/errors");
 
 module.exports.getUser = catchAsync(async (req, res) => {
+  if (!req.user.isAdmin) throw new ForbiddenError()
   const user = await User.findById(req.params.userId);
 
   res.status(StatusCodes.OK).json({ user });
 });
 
 module.exports.updateUser = catchAsync(async (req, res) => {
-  const validBody = await validateReqBodyFields(
-    ALLOWED_USER_FIELDS,
-    req.body
-  );
-  
+  if (!req.user.isAdmin) throw new ForbiddenError()
+
   await User.findByIdAndUpdate(req.params.userId,
-    { $set: validBody },
+    { $set: req.body },
     { runValidators: true });
 });
 
 module.exports.deleteUser = catchAsync(async (req, res) => {
+  if (!req.user.isAdmin) throw new ForbiddenError()
+
   await User.findByIdAndDelete(req.params.userId);
 });
